ref: Move shared schema code out of ClickHouse module
This change moves the shared column definitions from
`snuba.clickkhouse.columns` to `snuba.utils.schemas`. We will soon
be using these same column definitions for entities as well, and
we do not want to depend on the ClickHouse module as it is likely
to cause circular dependency issues.

The majority of this code has not changed, it is just copy and paste.

The ClickHouse column set now inherits from a base column set (which
currently does nothing). The same base class will also be used by
the entity column set class in the future.
